% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backprop_r.R
\name{backprop_gru}
\alias{backprop_gru}
\title{backprop_gru}
\usage{
backprop_gru(model, a, c, j, ...)
}
\arguments{
\item{model}{the output model object}

\item{a}{the input of this learning batch}

\item{c}{the output of this learning batch}

\item{j}{the indexes of the sample in the current batch}

\item{...}{argument to be passed to method}
}
\value{
the updated model
}
\description{
backpropagate the error in a model object of type gru
}

