% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backprop_r.R
\name{backprop_rnn}
\alias{backprop_rnn}
\title{backprop_rnn}
\usage{
backprop_rnn(model, a, c, j, ...)
}
\arguments{
\item{model}{the output model object}

\item{a}{the input of this learning batch, only one sample for the moment so batch is a little bit overrated}

\item{...}{argument to be passed to method}
}
\value{
the updated model
}
\description{
backpropagate the error in a model object of type rnn
}

