% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_r.R
\name{clean_lstm}
\alias{clean_lstm}
\title{clean_lstm}
\usage{
clean_lstm(model)
}
\arguments{
\item{model}{the output model object}
}
\value{
the updated model
}
\description{
clean the model for lighter output
}

