% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch_functions.R
\name{epoch_annealing}
\alias{epoch_annealing}
\title{epoch annealing}
\usage{
epoch_annealing(model)
}
\arguments{
\item{model}{the output model object}
}
\value{
the updated model
}
\description{
Apply the learning rate decay to the learning rate, called in epoch_model_function
}
