% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch_functions.R
\name{epoch_print}
\alias{epoch_print}
\title{epoch printing for trainr}
\usage{
epoch_print(model)
}
\arguments{
\item{model}{the output model object}
}
\value{
nothing
}
\description{
Print the error adn learning rate at each epoch of the trainr learning, called in epoch_function
}

