% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_r.R
\name{init_gru}
\alias{init_gru}
\title{init_gru}
\usage{
init_gru(model)
}
\arguments{
\item{model}{the output model object}
}
\value{
the updated model
}
\description{
Initialize the weight parameter for a gru
}

