% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_r.R
\name{init_lstm}
\alias{init_lstm}
\title{init_lstm}
\usage{
init_lstm(model)
}
\arguments{
\item{model}{the output model object}
}
\value{
the updated model
}
\description{
Initialize the weight parameter for a lstm
}

