% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_r.R
\name{init_rnn}
\alias{init_rnn}
\title{init_rnn}
\usage{
init_rnn(model)
}
\arguments{
\item{model}{the output model object}
}
\value{
the updated model
}
\description{
Initialize the weight parameter for a rnn
}

