% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int2bin.R
\name{int2bin}
\alias{i2b}
\alias{int2bin}
\title{Integer to Binary}
\usage{
int2bin(integer, length)

i2b(integer, length)
}
\arguments{
\item{integer}{input integer}

\item{length}{binary representation length}
}
\value{
binary representation
}
\description{
Integer to Binary
}
\section{Functions}{
\itemize{
\item \code{i2b}: individual Integer to Binary
}}

