% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int2binary.R
\name{int2binary}
\alias{int2binary}
\title{Integer 2 Binary}
\usage{
int2binary(x, length)
}
\arguments{
\item{x}{integer}

\item{length}{length of binary representation}
}
\description{
Converts an integer to binary form.
}

