% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{loss_L1}
\alias{loss_L1}
\title{L1 loss}
\usage{
loss_L1(model)
}
\arguments{
\item{model}{the output model object}
}
\value{
the updated model
}
\description{
Apply the learning rate to the weight update, vocabulary to verify !!
}

