% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictr.R
\name{predict_gru}
\alias{predict_gru}
\title{gru prediction function}
\usage{
predict_gru(model, X, hidden = FALSE, real_output = T, ...)
}
\arguments{
\item{model}{output of the trainr function}

\item{X}{array of input values, dim 1: samples, dim 2: time, dim 3: variables (could be 1 or more, if a matrix, will be coerce to array)}

\item{hidden}{should the function output the hidden units states}

\item{real_output}{option used when the function in called inside trainr, do not drop factor for 2 dimension array output}

\item{...}{arguments to pass on to sigmoid function}
}
\value{
array or matrix of predicted values
}
\description{
predict the output of a gru model
}
