% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictr.R
\name{predictr}
\alias{predictr}
\title{Recurrent Neural Network}
\usage{
predictr(model, X1, X2, binary_dim, alpha, input_dim, hidden_dim, output_dim,
  print = c("none", "minimal", "full"))
}
\arguments{
\item{model}{output of the trainr function}

\item{X1}{vector of input values}

\item{X2}{vector of input values}

\item{binary_dim}{dimension of binary representation}

\item{alpha}{size of alpha}

\item{input_dim}{dimension of input layer, i.e. how many numbers to sum}

\item{hidden_dim}{dimension of hidden layer}

\item{output_dim}{dimension of output layer}

\item{print}{should train progress be printed}
}
\value{
vector of predicted values
}
\description{
Trains a Recurrent Neural Network.
}
\examples{
# create training inputs
X1 = sample(0:127, 7000, replace=TRUE)
X2 = sample(0:127, 7000, replace=TRUE)

# create training output
Y <- X1 + X2

# train the model
m1 <- trainr(Y,
             X1,
             X2,
             binary_dim =  8,
             alpha      =  0.1,
             input_dim  =  2,
             hidden_dim = 10,
             output_dim =  1   )
             
# create test inputs
A1 = int2bin(sample(0:127, 7000, replace=TRUE), length=8)
A2 = int2bin(sample(0:127, 7000, replace=TRUE), length=8)
    
# predict
B  <- predictr(m1,
               A1,
               A2,
               binary_dim =  8,
               alpha      =  0.1,
               input_dim  =  2,
               hidden_dim = 10,
               output_dim =  1   )
 
# convert back to integers
A1 <- bin2int(A1)
A2 <- bin2int(A2)
B  <- bin2int(B)
 
# inspect the differences              
table( B-(A1+A2) )

# plot the difference
hist( B-(A1+A2) )

}

