% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnn.R
\name{rnn}
\alias{rnn}
\title{Recurrent Neural Network}
\usage{
rnn(binary_dim, alpha, input_dim, hidden_dim, output_dim, iterations = 5000)
}
\arguments{
\item{binary_dim}{dimension of binary representation}

\item{alpha}{size of alpha}

\item{input_dim}{dimension of input layer, i.e. how many numbers to sum}

\item{hidden_dim}{dimension of hidden layer}

\item{output_dim}{dimension of output layer}

\item{iterations}{number of training iterations}
}
\description{
Trains a Recurrent Neural Network.
}
\examples{
# training dataset generation
binary_dim = 8

# input variables
alpha = 0.1
input_dim = 2
hidden_dim = 16
output_dim = 1

# run
rnn(binary_dim, alpha, input_dim, hidden_dim, output_dim)
}

