% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.finance_demo.R
\name{run.finance_demo}
\alias{run.finance_demo}
\title{Demo app on finance data}
\usage{
run.finance_demo(port = NULL)
}
\arguments{
\item{port}{if not NULL will deploy on the local network on this port, just look for the IP of the host and go for example on 192.168.1.20:port to access the app from anywhere in the network}
}
\description{
Function to launch the finance_demo app
}
\examples{
\dontrun{
run.finance_demo()
}
}
\author{
Dimitri Fichou
}
