% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.rnn_demo.R
\name{run.rnn_demo}
\alias{run.rnn_demo}
\title{Demo app}
\usage{
run.rnn_demo(port = NULL)
}
\arguments{
\item{port}{if not NULL will deploy on the local network on this port, just look for the IP of the host and go for example on 192.168.1.20:port to access the app from anywhere in the network}
}
\description{
Function to launch the rnn_demo app
}
\examples{
\dontrun{
run.rnn_demo()
}
}
\author{
Dimitri Fichou
}

