% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid_output_to_derivative}
\alias{sigmoid_output_to_derivative}
\title{Sigmoid Derivative}
\usage{
sigmoid_output_to_derivative(output)
}
\arguments{
\item{output}{sigmoid value}
}
\description{
Convert output of sigmoid function to its derivative.
}

