% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainr.R
\name{trainr}
\alias{trainr}
\title{Recurrent Neural Network}
\usage{
trainr(Y, X, learningrate, learningrate_decay = 1, momentum = 0,
  hidden_dim = c(10), network_type = "rnn", numepochs = 1,
  sigmoid = c("logistic", "Gompertz", "tanh"), start_from_end = FALSE,
  use_bias = F, batch_size = 1, epoch_function = c(epoch_print),
  epoch_model_function = c(epoch_annealing), loss_function = loss_L1, ...)
}
\arguments{
\item{Y}{array of output values, dim 1: samples (must be equal to dim 1 of X), dim 2: time (must be equal to dim 2 of X), dim 3: variables (could be 1 or more, if a matrix, will be coerce to array)}

\item{X}{array of input values, dim 1: samples, dim 2: time, dim 3: variables (could be 1 or more, if a matrix, will be coerce to array)}

\item{learningrate}{learning rate to be applied for weight iteration}

\item{learningrate_decay}{coefficient to apply to the learning rate at each epoch, via the epoch_annealing function}

\item{momentum}{coefficient of the last weight iteration to keep for faster learning}

\item{hidden_dim}{dimension(s) of hidden layer(s)}

\item{network_type}{type of network, could be rnn or lstm, only rnn supported for the moment}

\item{numepochs}{number of iteration, i.e. number of time the whole dataset is presented to the network}

\item{sigmoid}{method to be passed to the sigmoid function}

\item{start_from_end}{should the sequence start from the end, legacy of the binary example. DEPRECATED, first index is the begining}

\item{use_bias}{should the network use bias}

\item{batch_size}{batch size: number of samples used at each weight iteration, only 1 supported for the moment}

\item{epoch_function}{vector of functions with no side effect on the model to applied at each epoch loop}

\item{epoch_model_function}{vector of functions with side effect on the model to applied at each epoch loop}

\item{loss_function}{loss function, applied in each sample loop, vocabulary to verify}

\item{...}{Arguments to be passed to methods, to be used in user defined functions}
}
\value{
a model to be used by the predictr function
}
\description{
Trains a Recurrent Neural Network.
}
\examples{
# create training numbers
X1 = sample(0:127, 7000, replace=TRUE)
X2 = sample(0:127, 7000, replace=TRUE)

# create training response numbers
Y <- X1 + X2

# convert to binary
X1 <- int2bin(X1, length=8)
X2 <- int2bin(X2, length=8)
Y  <- int2bin(Y,  length=8)

# create 3d array: dim 1: samples; dim 2: time; dim 3: variables
X <- array( c(X1,X2), dim=c(dim(X1),2) )

# train the model
model <- trainr(Y=Y,
                X=X,
                learningrate   =  0.1,
                hidden_dim     = 10,
                start_from_end = TRUE )
    
}

