% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_r.R
\name{update_adagrad}
\alias{update_adagrad}
\title{update_adagrad}
\usage{
update_adagrad(model)
}
\arguments{
\item{model}{the output model object}
}
\value{
the updated model
}
\description{
Apply the update with adagrad, not working yet
}
