% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_r.R
\name{update_lstm}
\alias{update_lstm}
\title{update_lstm}
\usage{
update_lstm(model)
}
\arguments{
\item{model}{the output model object}
}
\value{
the updated model
}
\description{
Apply the update for a lstm
}

