% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_r.R
\name{update_sgd}
\alias{update_sgd}
\title{update_sgd}
\usage{
update_sgd(model)
}
\arguments{
\item{model}{the output model object}
}
\value{
the updated model
}
\description{
Apply the update with stochastic gradient descent
}

