% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin2int.R
\name{bin2int}
\alias{b2i}
\alias{bin2int}
\title{Binary to Integer}
\usage{
bin2int(binary)

b2i(binary)
}
\arguments{
\item{binary}{input binary}
}
\value{
integer representation
}
\description{
Binary to Integer
}
\section{Functions}{
\itemize{
\item \code{b2i}: individual Binary to Integer
}}

