% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int2bin.R
\name{int2bin}
\alias{int2bin}
\title{Integer to Binary}
\usage{
int2bin(integer, length)
}
\arguments{
\item{integer}{input integer}

\item{output}{binary representation length}
}
\value{
binary representation
}
\description{
Integer to Binary
}

