% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_lstm.R
\name{predict_lstm}
\alias{predict_lstm}
\title{Long Short-Term Memory}
\usage{
predict_lstm(model, X, hidden = FALSE)
}
\arguments{
\item{model}{output of the train_lstm function}

\item{X}{array of input values, dim 1: samples, dim 2: time, dim 3: variables (could be 1 or more, if a matrix, will be coerce to array)}

\item{hidden}{should the function output the hidden units states}
}
\value{
array or matrix of predicted values
}
\description{
predict Long Short-Term Memory
}

