% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnn.R
\name{rnn}
\alias{rnn}
\title{Recurrent Neural Network}
\description{
A Recurrent Neural Network in native R, transforms numbers to binaries before adding bit by bit, teaching itself how to carry.
}
\seealso{
\code{\link{trainr}} for training a model and \code{\link{predictr}} for using a model to make predictions.
}
\author{
Bastiaan Quast \email{bquast@gmail.com}
}
