% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid}
\usage{
sigmoid(x, method = c("logistic", "SoftMax", "Gompertz"), inverse = FALSE,
  SoftMax = FALSE, ...)
}
\arguments{
\item{x}{number}
}
\description{
computes sigmoid nonlinearity
}

