% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainr.R
\name{trainr}
\alias{trainr}
\title{Recurrent Neural Network}
\usage{
trainr(Y, X1, X2, binary_dim, alpha, input_dim, hidden_dim, output_dim,
  print = c("none", "minimal", "full"))
}
\arguments{
\item{Y}{vector of output values}

\item{X1}{vector of input values}

\item{X2}{vector of input values}

\item{binary_dim}{dimension of binary representation}

\item{alpha}{size of alpha}

\item{input_dim}{dimension of input layer, i.e. how many numbers to sum}

\item{hidden_dim}{dimension of hidden layer}

\item{output_dim}{dimension of output layer}

\item{print}{should train progress be printed}
}
\value{
a model to be used by the predictr function
}
\description{
Trains a Recurrent Neural Network.
}
\examples{
# create training numbers
X1 = sample(0:127, 7000, replace=TRUE)
X2 = sample(0:127, 7000, replace=TRUE)

# create training response numbers
Y <- X1 + X2

# convert to binary
X1 <- int2bin(X1)
X2 <- int2bin(X2)
Y  <- int2bin(Y)

# train the model
trainr(Y,
       X1,
       X2,
       binary_dim =  8,
       alpha      =  0.1,
       input_dim  =  2,
       hidden_dim = 10,
       output_dim =  1,
       print = 'full'    )
    
}

